[#137] fix: 댓글 삭제 경고 추가

+++ src/features/club-detail/ui/comment-item.tsx
@@ -21,6 +21,8 @@ export default function CommentItem({
     commentId: number,
   ) => {
     e.preventDefault();
+    const confirmDelete = window.confirm('댓글을 삭제하시겠습니까?');
+    if (!confirmDelete) return;
 
     const response = await deleteComment(clubId, commentId);
     if (!response.ok) {
